+++ blog_Web/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class BlogWebConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'blog_Web'
