+++ blog_Web/models.py
@@ -1,0 +1,10 @@
+from django.db import models
+from django.utils import timezone
+
+class BlogPost(models.Model):
+    title = models.CharField(max_length=200)
+    content = models.TextField()
+    pub_date = models.DateTimeField('date published', default=timezone.now)
+
+    def __str__(self):
+        return self.title
